backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["IamIsthill <159409061+IamIsthill@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-split-settings = "^1.3.2"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 88
multi_line_output = 3
known_first_party = ["config"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
